modify

+++ lightning_attn_cute/test.py
@@ -122,10 +122,10 @@ def lightning_attn_no_decay(
     for i in range(NUM_BLOCK):
         si = i * BLOCK
         ei = min(si + BLOCK, N)
-        qi = q[:, :, si:ei].contiguous()
-        ki = k[:, :, si:ei].contiguous()
-        vi = v[:, :, si:ei].contiguous()
-        qkv_none_diag = torch.matmul(qi, kv.to(torch.float16))
+        qi = q[:, :, si:ei].contiguous().to(torch.float32)
+        ki = k[:, :, si:ei].contiguous().to(torch.float32)
+        vi = v[:, :, si:ei].contiguous().to(torch.float32)
+        qkv_none_diag = torch.matmul(qi, kv)
         o_inter_output[i] = qkv_none_diag.detach().clone()
 
 
@@ -140,7 +140,7 @@ def lightning_attn_no_decay(
 
         new_kv = torch.matmul(ki.transpose(-1, -2), vi)
         print(f"new_kv type: {new_kv.dtype}")
-        kv = kv + new_kv.to(torch.float32)
+        kv = kv + new_kv
         kv_output[i] = kv.detach().clone()
         print(f"data types. qi : {qi.dtype}, ki : {ki.dtype}, vi : {vi.dtype}, qkv_none_diag : {qkv_none_diag.dtype}, qk : {qk.dtype}, qkv_diag: {qkv_diag.dtype}, "
               f"output: {output.dtype}, new_kv: {new_kv.dtype}")
